Abstracted `PublicLinkScreen`
The abstracted screen will be used in place of `ShareCredentialScreen`.
