+++ app/navigation/CredentialNavigation/CredentialNavigation.d.tsx
@@ -1,5 +1,6 @@
 import type { StackScreenProps } from '@react-navigation/stack';
 import type { CredentialRecordRaw } from '../../model/credential';
+import { PublicLinkScreenParams } from '../../screens';
 
 export type CredentialNavigationParamList = {
   HomeScreen: undefined;
@@ -10,8 +11,10 @@ export type CredentialNavigationParamList = {
   ShareCredentialScreen: {
     rawCredentialRecord: CredentialRecordRaw
   };
+  PublicLinkScreen: PublicLinkScreenParams,
 };
 
 export type HomeScreenProps = StackScreenProps<CredentialNavigationParamList, 'HomeScreen'>;
 export type CredentialScreenHomeProps = StackScreenProps<CredentialNavigationParamList, 'CredentialScreen'>;
 export type ShareCredentialScreenProps = StackScreenProps<CredentialNavigationParamList, 'ShareCredentialScreen'>;
+export type PublicLinkScreenCredentialProps = StackScreenProps<CredentialNavigationParamList, 'PublicLinkScreen'>;
