+++ app/navigation/CredentialNavigation/CredentialNavigation.tsx
@@ -2,9 +2,8 @@ import React from 'react';
 
 import { createStackNavigator } from '@react-navigation/stack';
 
-import { HomeScreen, CredentialScreen } from '../../screens';
+import { HomeScreen, CredentialScreen, PublicLinkScreen } from '../../screens';
 import { CredentialNavigationProps, CredentialNavigationParamList } from '../';
-import { ShareCredentialScreen } from '../../screens';
 import { useResetNavigationOnBlur } from '../../hooks';
 
 const Stack = createStackNavigator<CredentialNavigationParamList>();
@@ -16,7 +15,7 @@ export default function CredentialNavigation({ navigation }: CredentialNavigatio
     <Stack.Navigator screenOptions={{ headerShown: false }}>
       <Stack.Screen name="HomeScreen" component={HomeScreen} />
       <Stack.Screen name="CredentialScreen" component={CredentialScreen} />
-      <Stack.Screen name="ShareCredentialScreen" component={ShareCredentialScreen} />
+      <Stack.Screen name="PublicLinkScreen" component={PublicLinkScreen} />
     </Stack.Navigator>
   );
 }
