+++ app/navigation/ShareNavigation/ShareNavigation.d.ts
@@ -1,6 +1,7 @@
 import type { StackScreenProps } from '@react-navigation/stack';
 
 import type { CredentialRecordRaw } from '../../model/credential';
+import { PublicLinkScreenParams } from '../../screens';
 
 export type ShareNavigationParamsList = {
   ShareHomeScreen: undefined;
@@ -14,13 +15,11 @@ export type ShareNavigationParamsList = {
   ShareSelectionScreen: {
     method: string
   };
-  PublicLinkScreen: {
-    rawCredentialRecord: CredentialRecordRaw;
-  }
+  PublicLinkScreen: PublicLinkScreenParams
 };
 
 export type ShareHomeScreenProps = StackScreenProps<ShareNavigationParamsList, 'ShareHomeScreen'>;
 export type PresentationPreviewScreenProps = StackScreenProps<ShareNavigationParamsList, 'PresentationPreviewScreen'>;
 export type CredentialScreenShareProps = StackScreenProps<ShareNavigationParamsList, 'CredentialScreen'>;
 export type ShareSelectionScreenProps = StackScreenProps<ShareNavigationParamsList, 'ShareSelectionScreen'>;
-export type PublicLinkScreenProps = StackScreenProps<ShareNavigationParamsList, 'PublicLinkScreen'>;
+export type PublicLinkScreenShareProps = StackScreenProps<ShareNavigationParamsList, 'PublicLinkScreen'>;
