+++ app/screens/PublicLinkScreen/PublicLinkScreen.d.ts
@@ -1,0 +1,7 @@
+import { PublicLinkScreenMode } from './PublicLinkScreen';
+export { PublicLinkScreenProps } from '../../navigation';
+
+export type PublicLinkScreenParams = {
+  rawCredentialRecord: CredentialRecordRaw;
+  screenMode?: PublicLinkScreenMode;
+};
